fix: address copilot review and self-review findings
Doc + UX fixes from copilot review on PR #158:

- docs/agents.md: 3 broken markdown links (src/... resolved relative
  to docs/, producing docs/src/...). Switched to ../src/... paths.

- README.md: dropped specific Antigravity model lineup ("Gemini 3,
  Claude, GPT-OSS"), per the same neutrality principle as the onboard
  hints. Antigravity's lineup may shift; users discover models via
  config rather than promotional copy.

- src/shared/agent-models.ts: moved validateAgentModels (and its
  helpers) here from src/cli/utils/config.ts, so the server's
  AgentHost no longer imports across the cli/server boundary. Added
  per-process memoization of getProviders()/getModels() so validation
  across N agents and M reinitialization paths walks pi-ai's catalog
  at most once per provider. Cli/utils/config.ts now re-exports
  validateAgentModels from shared so existing CLI consumers and tests
  keep working unchanged.

- src/cli/utils/config.ts: buildConfigToml now emits a comment block
  above [llm], [llm.oauth], and each [llm.<provider>] section, so a
  user reading their own config.toml understands the two-tier model,
  the supported OAuth provider IDs, where tokens live, and how to
  manage them. Auth sections were previously the only undocumented
  area of the file.

- config.test.ts: tightened the "omits oauth section when undefined"
  assertion to anchor on a section header (^[llm.oauth]) at line
  start, since the new comment block legitimately mentions
  "[llm.oauth]" in prose without emitting the section.

Skipped (per user direction): rebuilding fallback options from the
full provider list when a primary OAuth login fails. Current behavior
keeps the failed candidate out of the fallback selection too — a
trade-off the user is comfortable with.

873 tests pass. lint/build clean.
